+++ Content.Shared/RCD/Components/RCDAmmoComponent.cs
@@ -12,5 +12,5 @@ public sealed partial class RCDAmmoComponent : Component
     /// Can be partially transferred into an RCD, until it is empty then it gets deleted.
     /// </summary>
     [DataField("charges"), ViewVariables(VVAccess.ReadWrite), AutoNetworkedField]
-    public int Charges = 30;
+    public float Charges = 30;
 }
