+++ Content.Shared/RCD/RCDPrototype.cs
@@ -48,7 +48,7 @@ public sealed partial class RCDPrototype : IPrototype
     /// Number of charges consumed when the operation is completed
     /// </summary>
     [DataField, ViewVariables(VVAccess.ReadOnly)]
-    public int Cost { get; private set; } = 1;
+    public float Cost { get; private set; } = 1;
 
     /// <summary>
     /// The length of the operation
